I agree with marking out priests, but not this way
I removed this check, because as the game goes on, these priest dwarf's
WILL fall in skill compared to other dwarfs and already be the lower
picked for any role.  And some priests I have found in game to already
have skills, so they may actually be better than other dwarf's ATM.  So
instead, let the skill rate formula do it's job in this calculation.